refactor and use proper namespace
