+++ markdown/coding/src/challenges/refactor/Article.java
@@ -1,4 +1,4 @@
-package challenges;
+package challenges.refactor;
 
 public class Article {
