+++ markdown/coding/src/challenges/refactor/Bike.java
@@ -1,4 +1,4 @@
-package challenges;
+package challenges.refactor;
 
 public class Bike {
