+++ markdown/coding/src/challenges/refactor/Brompton.java
@@ -1,22 +1,26 @@
-package challenges;
+package challenges.refactor;
 
-public class EBike extends Bike {
+public class Brompton extends Bike {
 
     public int maxSpeed;
     public int rearGearsCount;
     public int frontGearsCount;
 
-    public EBike(String pn, double p, int ms, int rgc, int fgc, int bc) {
+    public Brompton(String pn, double p, int ms, int rgc, int fgc) {
         productName = pn;
         price = p;
         maxSpeed = ms;
         rearGearsCount = rgc;
         frontGearsCount = fgc;
-        batteryCapacity = bc;
     }
 
     public int getMaxSpeed() {
         return maxSpeed;
+    }
+
+    @Override
+    public Integer getBatteryCapacity() {
+        return null;
     }
 
     @Override
