+++ markdown/coding/src/challenges/refactor/Mountainbike.java
@@ -1,4 +1,4 @@
-package challenges;
+package challenges.refactor;
 
 public class Mountainbike extends Bike {
